+++ vkconfig_core/test/test_executable.cpp
@@ -131,25 +131,25 @@ TEST(test_executable, GetDefaultExecutablePath) {
     qunsetenv("VULKAN_SDK");
 
 #if defined(_WIN32)
-    DefaultExecutable default_executable{"vkcube.exe", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    DefaultExecutable default_executable{"vkcube.exe", "vkcube", {{"vkcube launcher options", "", "--suppress_popups"}}};
     const DefaultPath& default_paths = ::GetDefaultExecutablePath(default_executable.key);
     EXPECT_STREQ(default_paths.executable_path.RelativePath().c_str(), Path("vkcube.exe").RelativePath().c_str());
 #elif defined(__linux__)
-    DefaultExecutable default_executable{"vkcube", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    DefaultExecutable default_executable{"vkcube", "vkcube", {{"vkcube launcher options", "", "--suppress_popups"}}};
     const DefaultPath& default_paths = ::GetDefaultExecutablePath(default_executable.key);
     EXPECT_STREQ(default_paths.executable_path.RelativePath().c_str(), Path("vkcube").RelativePath().c_str());
 #elif defined(__APPLE__)
-    DefaultExecutable default_executable{"vkcube.app", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    DefaultExecutable default_executable{"vkcube.app", "vkcube", {{"vkcube launcher options", "", "--suppress_popups"}}};
     const DefaultPath& default_paths = ::GetDefaultExecutablePath(default_executable.key);
-    // EXPECT_STREQ(default_paths.executable_path.RelativePath().c_str(),
-    // Path("/Applications/vkcube.app/Contents/MacOS/vkcube").RelativePath().c_str());
+    EXPECT_STREQ(default_paths.executable_path.RelativePath().c_str(),
+                 Path("/Applications/vkcube.app/Contents/MacOS/vkcube").RelativePath().c_str());
 #endif
 
 #if defined(_WIN32) || defined(__linux__)
     EXPECT_TRUE(default_paths.working_folder.Empty());
 #elif defined(__APPLE__)
-    // EXPECT_STREQ(default_paths.working_folder.RelativePath().c_str(),
-    // Path("/Applications/vkcube.app/Contents/MacOS").RelativePath().c_str());
+    EXPECT_STREQ(default_paths.working_folder.RelativePath().c_str(),
+                 Path("/Applications/vkcube.app/Contents/MacOS").RelativePath().c_str());
 #endif
 
     qputenv("VULKAN_SDK", saved.c_str());
@@ -160,18 +160,18 @@ TEST(test_executable, DefaultExecutable) {
     qunsetenv("VULKAN_SDK");
 
 #if defined(_WIN32)
-    DefaultExecutable default_executable{"vkcube.exe", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    DefaultExecutable default_executable{"vkcube.exe", "vkcube", {{"vkcube launcher options", "", "--suppress_popups"}}};
     Executable executable(default_executable);
     EXPECT_STREQ(executable.path.RelativePath().c_str(), Path("vkcube.exe").RelativePath().c_str());
 #elif defined(__linux__)
-    DefaultExecutable default_executable{"vkcube", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    DefaultExecutable default_executable{"vkcube", "vkcube", {{"vkcube launcher options", "", "--suppress_popups"}}};
     Executable executable(default_executable);
     EXPECT_STREQ(executable.path.RelativePath().c_str(), Path("vkcube").RelativePath().c_str());
 #elif defined(__APPLE__)
-    DefaultExecutable default_executable{"vkcube.app", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    DefaultExecutable default_executable{"vkcube.app", "vkcube", {{"vkcube launcher options", "", "--suppress_popups"}}};
     Executable executable(default_executable);
-    // EXPECT_STREQ(executable.path.RelativePath().c_str(),
-    // Path("/Applications/vkcube.app/Contents/MacOS/vkcube").RelativePath().c_str());
+    EXPECT_STREQ(executable.path.RelativePath().c_str(),
+                 Path("/Applications/vkcube.app/Contents/MacOS/vkcube").RelativePath().c_str());
 #endif
 
     EXPECT_EQ(1, executable.GetOptions().size());
@@ -182,8 +182,8 @@ TEST(test_executable, DefaultExecutable) {
 #if defined(_WIN32) || defined(__linux__)
     EXPECT_TRUE(options->working_folder.Empty());
 #elif defined(__APPLE__)
-    // EXPECT_STREQ(options->working_folder.RelativePath().c_str(),
-    // Path("/Applications/vkcube.app/Contents/MacOS").RelativePath().c_str());
+    EXPECT_STREQ(options->working_folder.RelativePath().c_str(),
+                 Path("/Applications/vkcube.app/Contents/MacOS").RelativePath().c_str());
 #endif
     EXPECT_STREQ(options->args[0].c_str(), "--suppress_popups");
     EXPECT_TRUE(options->envs.empty());